+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.1.2] / 2022-02-23
+### Bug Fixes
+- Null GetFileVersionInfo set Warning
 ### Changed
+- Add GetAssemblyName
 - Add GetComments & GetFileDescription
 - Add ShowInformation On MainProject